payment amount 단위 수정, entity setter 제거, Payment TDD 구성

+++ backend/src/main/java/com/caffe/global/component/MockPaymentGatewayClient.java
@@ -5,8 +5,8 @@ import org.springframework.stereotype.Component;
 @Component
 public class MockPaymentGatewayClient {
 
-    public boolean charge(String paymentMethod, String paymentName,  String paymentInfo, double amount) {
-        System.out.printf("결제 방법 : %s , 결제사 : %s, 결제 요청 번호 : %s, 결제 금액 : %f", paymentMethod, paymentName, paymentInfo, amount);
+    public boolean charge(String paymentMethod, String paymentName,  String paymentInfo, int amount) {
+        System.out.printf("결제 방법 : %s , 결제사 : %s, 결제 요청 번호 : %s, 결제 금액 : %d", paymentMethod, paymentName, paymentInfo, amount);
         try{
             Thread.sleep(500);
         }catch (InterruptedException e){
